feat: static gen sitemap and rss
